pyproject.toml
[project]
name = "config_reader"
description = "Simple Yaml config reader"
version = "0.1.1"

requires-python = ">= 3.10"
classifiers = [
    "Development Status :: 0 - Pre Alpha",
    "Programming Language :: Python"
]

dependencies = [
    "pyyaml",
    "jmespath",
    "python-dotenv"
]


[tool.setuptools]
packages = [
    "config_reader",
    "project_config"
]


[build-system]
requires = [
    "setuptools @ git+https://github.com/pypa/setuptools",
    "cython >= 0.29.33",
    "wheel >= 0.40"
]
build-backend = "setuptools.build_meta"

[project.scripts]
    init = 'project_config.setup:main'

#[tool.setuptools.package-data]
#package_name = [
#    "*.h5", "*.txt"
#]

#[project.scripts]
#    my-script = "my_package.module:function"